refactor: extract Claude session reconciliation
